+++ docs/examples/02_hello_threaded_world.py
@@ -1,26 +1,25 @@
 import time
 from threading import Thread
 
-from compas_eve import Message
 from compas_eve import Publisher
 from compas_eve import Subscriber
 from compas_eve import Topic
 
-topic = Topic("/compas_eve/hello_world/", Message)
+topic = Topic("/compas_eve/hello_world/")
 
 
 def start_publisher():
     publisher = Publisher(topic)
 
     for i in range(20):
-        msg = Message(text=f"Hello world #{i}")
-        print(f"Publishing message: {msg.text}")
+        msg = dict(text=f"Hello world #{i}")
+        print(f"Publishing message: {msg}")
         publisher.publish(msg)
         time.sleep(1)
 
 
 def start_subscriber():
-    subcriber = Subscriber(topic, callback=lambda msg: print(f"Received message: {msg.text}"))
+    subcriber = Subscriber(topic, callback=lambda msg: print(f"Received message: {msg}"))
     subcriber.subscribe()